pyproject.toml
[project]
name = "vibecut"
version = "0.1.0"
description = "AI-powered video clip production using agent skills"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "vibecut contributors"}
]
keywords = ["video", "ai", "remotion", "voice-cloning", "captions"]

# Core dependencies - minimal, most skills work without these
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
# Video analysis (Gemini)
video = ["google-genai>=1.0.0"]
# ASR and caption alignment
asr = ["qwen-asr>=0.1.0", "jieba>=0.42.1"]
# Audio processing
audio = ["soundfile>=0.12.0", "scipy>=1.10.0", "pydub>=0.25.1"]
# All features
all = ["vibecut[video,asr,audio]"]

[project.urls]
Homepage = "https://github.com/yourusername/vibecut"
Repository = "https://github.com/yourusername/vibecut.git"

[tool.uv]
dev-dependencies = ["pytest>=8.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
